benchmarking: add infrastructure and initial code
Add initial infrastructure code for declaring, registering, and running
benchmarks. Introduce an initial benchmark, pchase_cpu, that performs
cache size estimation for a CPU cache using the pointer-chasing
algorithm. It serves as an an example for how future benchmarks can be
declared.

The README is updated with some style guide changes etc...