cudaHelpers: update throwOnErr, add modulo_power_of_2
This patch updates throwOnErr to be a macro, allows it to be more
descriptive by including the file and line number in the error message
that it emitted.

Also adds a helper macro modulo_power_of_2, that uses bitwise arithmetic
for modulo computation which is more efficient than a % operator.