clock64.cu: add helper for reading maximum clock64 frequency
Adds a function, getMaxClockFrequencyHz(), that returns device 0's
maximum clock frequency in Hz. In particular, this is the SM clock
which corresponds to the frequency of the clock that clock64()
probes.

As this pulls in a new dependency for nvml, our workflows configuration
is updated, as well as the Makefile that must link against it.