global-to-shared: add benchmark for global to shared mem transfers
Introduce a new benchmark for measuring the bandwidth of transfers from
global memory (DRAM) to shared memory. We parameterize this with some
synthetic work performed on each element.

In particular, we test with two distinct read strategies:
- Synchronous reads, where all threads in a block collaborate in reading
  from DRAM with perfectly coalesced accesses, and write into shared
  memory without bank conflicts.
- Asynchronous reads with double buffering - reads are done
  asynchronousl with cuda::memcpy_async while the threads perform the
  synthetic work.

The idea is to measure how much work must be performed per element for
it to be worthwhile to use asynchronous copying, which itself introduces
some overhead due to barrier synchronization.

The results are promising, showing that asynchronous copies perform
better in compute-bound kernels as memory accesses can be made while
threads do their work. See the plot in `plots/` (this is just for
posterity, the plot isn't great and is poorly named).

Signed-off-by: Ethan Graham <ethan.w.s.graham@gmail.com>