Makefile: NVCC should not break for conversion errors
Our CI is failing due to conversion errors related to some CUDA header
file [1]. I am unable to reproduce this locally, and after trying to
bump the GCC version (in vain) in e92f6a3, I'm thinking that it's best
to just disable errors for this warning.

We also add up-to-date flags for the blackwell architecture. This should
make our code more efficient, but I guess we will just have to play
around with this. We remove the GENCODES lines because it was breaking
the build sometimes - it's probably best just to target the specific
arch of the blackwell GPUs that we use on the student cluster.

[1] https://github.com/ethangraham2001/gpu-memperf/actions/runs/19414876628

Signed-off-by: Ethan Graham <ethan.w.s.graham@gmail.com>