benchmarks: use floats instead of unsigned ints
As discussed with Yakub, it is more intersting to use floating point
values for computation rather than integers. In GPU programming,
integers are generally just used for offset calculation, and for this
reason floating point units are usually more optimized for general
computation.

This patch proposes to keep the type alises for f32 and f64, but under
the hood replace them with real float values. This also required some
changes to the inline PTX for the L1 random access benchmark.