strided-access: refactor and use cuda events instead of clock64
There was a fundamental flaw in how we were measuring time for our
strided access benchmark. In particular, we were using per-block
clock64 timers, and then choosing the longest block as proxy for the
total number of cycles taken.

Blocks are the schedulable unit on CUDA GPUs, and we have no guarantee
that all blocks will run in lockstep. Assume the following:

- Block 0 starts at time S_0, accesses all data, finished at time E_0.
- Some time _after_ E_0, block 35 starts at time S_35 and ends at time
  E_35, with (E_35 - S_35) > (E_0 - S_0).

We then report a time of (E_35 - S35), when in reality it took longer.

To mitigate this error-prone behavior, we use wall-clock time instead,
and use a warmup loop. We know that data will not be evicted from cache
if it is associated to the same device address that hasn't been freed.

Signed-off-by: Ethan Graham <ethan.w.s.graham@gmail.com>