REVERT THIS
Adds util::randomVector manually (included in another unmerged PR). This
was to verify the hypothesis that we were exceeding the theoretical peak
bandwidth on our DRAM measurements due to potential data compression.
When using fully pseudo-random data (instead of all zeroes) we do indeed
observe lower bandwidth, in-line with expectations.