orchestrator: add orchestrator program for bulk-benchmarking
Up until now, we have been benchmarking programs individually. This is
error-prone, as configurations can diverge, as well as the current
version of the repository that is being used for the benchmarking.

This patch introduces a one-file Python orchestrator program that is
designed so that we can check in configurations (into the Python) file,
bulk run out benchmarks, and collect all plots in a single directory.

At a high-level, all it is doing is calling gpu-memperf.

Structurally, each benchmark is expected to implement some base class,
that tells the "orchestrator" how to call it, and how to plot its
results. The user can then specify which benchmarks to run, and then
benchmark them in sequence and aggregate their results.

Although it works pretty well for the global-to-shared and
shared-to-register benchmarks, some changes are still needed for the
other benchmarks, as their Python plotting code was modular enough to
just call into. However, the other benchmarks seem to be hard-coding
their results in the Python files directly. While this was fine for
initial results collection, it becomes a huge PITA as soon as we collect
more results / want to run them many times and compare - these will
require some updates.